// app/web/src/data/Model/AppError.ts
export enum AppError {
    NetworkError = "NetworkError",
    AppOutdated = "AppOutdated",
    InvalidResponse = "InvalidResponse",
    InvalidUsername = "InvalidUsername",
    UsernameTaken = "UsernameTaken",
    Unauthorized = "Unauthorized",
    ServerUpdating = "ServerUpdating",
    ServerError = "ServerError",
    Unknown = "Unknown",
}